api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
sqlalchemy==2.0.35
pydantic==2.10.3
python-dotenv==1.0.1
email-validator==2.2.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.18
confluent-kafka==2.3.0
prometheus-fastapi-instrumentator==6.1.0
psycopg2-binary==2.9.10
